Refactor to have Searcher
