+++ backend/Assistant.py
@@ -6,102 +6,12 @@ from datetime import datetime
 from rich import print, table
 
 
-class assistant:
-    def __init__(self, openai_api_key, voyageai_api_key, db_uri):
+class Assistant:
+    def __init__(self, openai_api_key, searcher):
         print("[bold]Creating Chatbot[/bold]")
-        print(f"connecting to database at {db_uri}")
-        self.vector_db = lancedb.connect(db_uri)
-        table_name = "all_document_types"
-        self.all_document_types_table = self.vector_db.open_table(table_name)
+        self.searcher = searcher
         self.openai_client = openai.OpenAI(api_key=openai_api_key)
-        self.voyageai_client = voyageai.Client(api_key=voyageai_api_key)
-
-        print("[bold]Created Chatbot[/bold]")
-        chatbot_config = table.Table(title="🛠️  Chatbot Configuration 🛠️")
-
-        chatbot_config.add_column("Name")
-        chatbot_config.add_column("Value")
-        chatbot_config.add_row("Database URI", db_uri)
-        chatbot_config.add_row("Table Name", table_name)
-        chatbot_config.add_row(
-            "Table Size", f"{self.all_document_types_table.count_rows()} rows"
-        )
-        chatbot_config.add_row(
-            "Columns", ", ".join(self.all_document_types_table.schema.names)
-        )
-        print(chatbot_config)
-
-        if "agency" not in self.all_document_types_table.schema.names:
-            raise ValueError("agency column not found in table")
-
-    def embed_query(self, query: str):
-        return self.voyageai_client.embed(
-            query, model="voyage-large-2-instruct", input_type="query", truncation=False
-        ).embeddings[0]
-
-    def knowledge_search(
-        self,
-        query: str,
-        type: str,
-        year_range: tuple[int, int],
-        document_type: list[str],
-        modes: list[str],
-        agencies: list[str],
-    ):
-        limit = 100
-        where_statement = []
-        if year_range:
-            where_statement.append(
-                f"year >= {year_range[0]} and year <= {year_range[1]}"
-            )
-        if document_type:
-            document_types = ", ".join(f'"{dt}"' for dt in document_type)
-            where_statement.append(f"document_type in ({document_types})")
-        if modes and len(modes) > 1:
-            where_statement.append(f"mode in {tuple([str(mode) for mode in modes])}")
-        elif modes and len(modes) == 1:
-            where_statement.append(f"mode = '{modes[0]}'")
-        if agencies and len(agencies) > 1:
-            where_statement.append(f"agency in {tuple(agencies)}")
-        elif agencies and len(agencies) == 1:
-            where_statement.append(f"agency = '{agencies[0]}'")
-
-        where_statement = " AND ".join(where_statement)
-
-        if query == "" or query is None:
-            final_query = None
-        elif type == "fts":
-            final_query = query
-        elif type == "vector":
-            final_query = self.embed_query(query)
-        else:
-            raise ValueError(f"type must be 'fts' or 'vector' not {type}")
-
-        query_table = table.Table(
-            title="🔍 Conducting search with 🔍",
-            show_header=True,
-            title_style="bold blue",
-        )
-        query_table.add_column("Parameter")
-        query_table.add_column("Value")
-        query_table.add_row(
-            "Query",
-            final_query
-            if isinstance(final_query, str)
-            else "vector embeddings of " + query,
-        )
-        if where_statement:
-            query_table.add_row("Filters", where_statement)
-        print(query_table)
-
-        results = (
-            self.all_document_types_table.search(final_query, query_type=type)
-            .where(where_statement, prefilter=True)
-            .limit(limit)
-            .to_pandas()
-        ).drop(columns=["vector"])
-
-        return results
+        print("[bold]Chatbot created[/bold]")
 
     def provide_conversation_title(self, history=[]):
         if history == []:
@@ -129,13 +39,13 @@ class assistant:
         system_message = {
             "role": "system",
             "content": f"""
-You are a expert working at the New Zealand transport accident investigation commision. Your job is to assistant employees of TAIC with their queries. The day is {datetime.now()}.
+You are a expert working at the New Zealand transport accident investigation commission. Your job is to assistant employees of TAIC with their queries. The day is {datetime.now()}.
 You will be provided the conversation history and a query from the user. You can either respond directly or can call a function that searches a database of accident reports.
 For each report you should provide use the report IDs, if you reference any other document you should provide the document type and document ID.
 
 Here is some more dataset information
-There are {len(self.all_document_types_table.schema.names)} columns with {self.all_document_types_table.count_rows()} rows.
-The columns available are: {"".join(self.all_document_types_table.schema.names)}
+There are {len(self.searcher.all_document_types_table.schema.names)} columns with {self.searcher.all_document_types_table.count_rows()} rows.
+The columns available are: {"".join(self.searcher.all_document_types_table.schema.names)}
 
 Here are some definitions from TAIC:
 
@@ -233,8 +143,6 @@ Example function calls:
                                 "modes",
                                 "agencies",
                             ],
-                            "additionalProperties": False,
-                            "strict": True,
                         },
                     },
                 }
@@ -287,7 +195,7 @@ Example function calls:
         )
         yield history
 
-        results = self.knowledge_search(**function_arguments)
+        results = self.searcher.knowledge_search(**function_arguments)
         tool_call_message = {
             "role": "assistant",
             "content": None,
