+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "TAIC smart assistant"
 version = "0.1.0"
-description = "A demo app that uses a vector database of accident investigation reports from TAIC, ATSB and TSB to assist investigators and reearchers."
+description = "A demo app that uses a vector database of accident investigation reports from TAIC, ATSB and TSB to assist investigators and researchers."
 authors = ["1jamesthompson1 <1jamesthompson1@gmail.com>"]
 readme = "README.md"
 
@@ -23,13 +23,17 @@ itsdangerous = "^2.2.0"
 azure-data-tables = "^12.6.0"
 datasets = "^3.2.0"
 pylance = "^0.25.0"
+gradio-rangeslider = "^0.0.8"
+transformers = "^4.51.3"
+pydantic = "2.10.6"
 
 [tool.poetry.scripts]
 chatbot = "app:main"
 
 [tool.poetry.group.dev.dependencies]
-huggingface-hub = {extras = ["cli"], version = "^0.27.0"}
+huggingface-hub = {extras = ["cli"], version = ">=0.30.0,<1.0"}
 ipykernel = "^6.29.5"
+uvicorn = "^0.34.2"
 
 [build-system]
 requires = ["poetry-core"]
